Правки. Перечитал тритий пункт.
