+++ src/components/order-details/order-details.js
@@ -1,12 +1,14 @@
 import React from 'react';
 import modalImg from '../../images/modal-order.png';
-import PropTypes from 'prop-types';
 import style from './order-details.module.css';
+import { useSelector } from 'react-redux';
 
-export default function OrderDetails({ orderId }) {
+export default function OrderDetails() {
+    const order = useSelector(state => state.order);
+
     return (
             <div className={style.mainModal}>
-            <span className={`text text_type_digits-large pt-20 pb-8 ${style.modalIdOrder}`}>{orderId}</span>
+            <span className={`text text_type_digits-large pt-20 pb-8 ${style.modalIdOrder}`}>{order.orderDetails.order.number}</span>
                 <div className={`text text_type_main-medium pb-15 ${style.modalIdHint}`}>Идентификатор заказа</div>
                 <div className={`pb-15 ${style.modalImg}`}><img src={modalImg} /></div>
                 <div className={`text text_type_main-default pb-2  ${style.modalStatus}`}>Ваш заказ начали готовить</div>
@@ -14,7 +16,3 @@ export default function OrderDetails({ orderId }) {
             </div>
         );
 }
-
-OrderDetails.propTypes = {
-    orderId: PropTypes.number.isRequired
-};
