+++ src/services/actions/burger-constructor.js
@@ -1,3 +1,3 @@
 export const ADD_TO_CONSTRUCTOR      = 'ADD_TO_CONSTRUCTOR';
 export const REMOVE_FROM_CONSTRUCTOR = 'REMOVE_FROM_CONSTRUCTOR';
-export const OVERWRITING_CONSTRUCTOR = 'OVERWRITING_CONSTRUCTOR';
+export const MOVE_IN_CONSTRUCTOR = 'MOVE_IN_CONSTRUCTOR';
